requirements.txt
stegano
Pillow
opencv-python
cryptography
pydub
ttkthemes
